requirements.txt
streamlit>=1.26
pandas
folium
streamlit-folium
python-dotenv
requests
tqdm